+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     * Fixed user avatar creation for new `Flex Users` when using folder storage
     * Fixed `Trying to access array offset on value of type null` PHP 7.4 error in `Plugin.php`
     * Fixed Gregwar Image library using `.jpeg` for cached images, rather use `.jpg`
+    * Implemented missing Flex `$pageCollection->batch()` and `$pageCollection->order()` methods
 
 # v1.7.0-rc.8
 ## 03/19/2020